tpm: Implement a virtio driver
Implement a virtio driver that uses a queue with 4 elements. Two of the
elements are for sending a header and the TPM command and the other two
are for passing the receive buffer and a field for result status code.

Read the version of the TPM device from the virtio configuration space.

For power management, call tpm_pm_resume_state() so that the TPM can be
resume with the state that was saved upon freeze. It is necessary to resume
with the TPM Savestate since the QEMU board will be reset, which in turn
also resets the TPM device, requiring the restoring of the TPM Savestate.

Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>